requirements.txt
random-address
kafka-python